refactor: move logic to Pool page
