+++ src/views/Pool.vue
@@ -113,6 +113,10 @@ import { useI18n } from 'vue-i18n';
 import { useRoute, useRouter } from 'vue-router';
 import useNumbers from '@/composables/useNumbers';
 import { getTokensHistoricalPrice, HistoricalPrices } from '@/api/coingecko';
+import { useQueryClient } from 'vue-query';
+
+import { POOLS_ROOT_KEY } from '@/constants/queryKeys';
+
 import {
   getPoolEvents,
   getPoolSnapshots,
@@ -132,7 +136,10 @@ interface PoolPageData {
   events: PoolEvents;
   prices: HistoricalPrices;
   snapshots: PoolSnapshots;
+  refetchQueriesOnBlockNumber: number;
 }
+
+const REFETCH_QUERIES_BLOCK_BUFFER = 3;
 
 export default defineComponent({
   components: {
@@ -155,6 +162,7 @@ export default defineComponent({
       blockNumber,
       loading: web3Loading
     } = useWeb3();
+    const queryClient = useQueryClient();
 
     // DATA
     const data = reactive<PoolPageData>({
@@ -166,7 +174,8 @@ export default defineComponent({
         exits: []
       },
       prices: {},
-      snapshots: []
+      snapshots: [],
+      refetchQueriesOnBlockNumber: 0
     });
 
     // COMPUTED
@@ -230,6 +239,9 @@ export default defineComponent({
         pool.value.address
       ]);
       console.timeEnd('loadPool');
+
+      data.refetchQueriesOnBlockNumber =
+        blockNumber.value + REFETCH_QUERIES_BLOCK_BUFFER;
     }
 
     async function loadEvents(): Promise<void> {
@@ -257,6 +269,9 @@ export default defineComponent({
         await fetchPool();
         await loadEvents();
         data.backgroundLoading = false;
+      }
+      if (data.refetchQueriesOnBlockNumber === blockNumber.value) {
+        queryClient.invalidateQueries([POOLS_ROOT_KEY]);
       }
     });
